+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [5.1.0] (IN PROGRESS)
 
 * Update the .gitignore file. Refs UICHKIN-232.
+* Add pull request template. Refs UICHKIN-233.
 
 ## [5.0.0] (https://github.com/folio-org/ui-checkin/tree/v5.0.0) (2021-03-10)
 [Full Changelog](https://github.com/folio-org/ui-checkin/compare/v4.0.2...v5.0.0)