+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Update the .gitignore file. Refs UICHKIN-232.
 * Add pull request template. Refs UICHKIN-233.
+* Navigating back to `Checkin` page leads to error message. Refs UICHKIN-239.
 * Add settings up for Jest/RTL tests. Refs UICHKIN-237.
 
 ## [5.0.0] (https://github.com/folio-org/ui-checkin/tree/v5.0.0) (2021-03-10)