+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Update the .gitignore file. Refs UICHKIN-232.
 * Add pull request template. Refs UICHKIN-233.
 * Navigating back to `Checkin` page leads to error message. Refs UICHKIN-239.
+* The fee/fine account is fully closed when Claimed Returned item is checked in. Refs UICHKIN-235.
 * Add settings up for Jest/RTL tests. Refs UICHKIN-237.
 
 ## [5.0.0] (https://github.com/folio-org/ui-checkin/tree/v5.0.0) (2021-03-10)